Split the library into its own file
